feat: Centralize error handling with errorMessages module
- Created errorMessages.js to manage consistent error messaging across the application.
- Added failing test cases for all bath-related controllers to ensure robust error handling.
- Optimized HTTP status codes in controllers for accurate client feedback.

+++ backend-sc/controllers/bath.controller.js
@@ -3,6 +3,7 @@
  * @module BathController
  * @description This module handles all CRUD operations related to baths.
  * @requires ../models - Database models needed for bath operations.
+ * @requires ../config.js/errorMessages - Error messages for different HTTP status codes.
  * @exports createBath - Creates a new Bath record in the database.
  * @exports modifyBath - Modifies an existing Bath record in the database.
  * @exports deleteBath - Delete an existing Bath record in the database.
@@ -17,7 +18,10 @@
  */
 
 // Import dependencies
+const mongoose = require("mongoose");
+const errorMessages = require("../config.js/errorMessages");
 const db = require("../models");
+const e = require("express");
 
 const Bath = db.bath;
 const User = db.user;
@@ -25,72 +29,109 @@ const User = db.user;
 // Define the sort options for bath records
 const sortOptions = { createdAt: -1 };
 
+// Define the required fields for bath records
+const requiredFields = [
+  "author",
+  "waterTemperature",
+  "timeInWater",
+  "temperatureOutside",
+  "weather",
+];
+
 /**
  * @function createBath
  * @async
- * @description Creates a new Bath record in the database.
+ * @description Creates a new Bath record in the database and returns the newly created record.
  * @see {@link module:BathRoutes} - This function is used in the POST /api/bath route.
  * @param {Object} req - Express request object, containing the Bath details in the body.
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message and the newly created bath.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.post("/api/bath", [authJwt.verifyToken], controller.createBath);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if a required field is missing.
+ * @throws {BadRequest} JSON response with a 400 status if the user ID is invalid.
+ * @throws {NotFound} JSON response with a 404 status if the user is not found.
+ * @example app.post("/api/bath", [authJwt.verifyToken], controller.createBath);
  */
 exports.createBath = async (req, res, next) => {
   try {
     // Remove the fake_id sent by the front-end
     delete req.body._id;
-
-    // Verify if the user ID exists
-    const userId = req.body.author;
-    const userExists = await User.findById(userId);
-
-    if (!userExists) {
-      // Return a generic error message or a misleading success message
-      return res.status(400).json({ message: "Une erreur s'est produite." });
+    // Verify if all required fields are present
+    for (const field of requiredFields) {
+      if (!req.body.hasOwnProperty(field)) {
+        console.log(`Missing field: ${field}`); // Debug log
+        return res.status(400).json({ message: "Champs requis manquants." });
+      }
     }
 
+    const userId = req.body.author;
+
+    // Check if userId is valid
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
+    // Verify if the user ID exists
+    const userExists = await User.findById(userId);
+    if (!userExists) {
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
+    }
     // Create a new bath record
     const bath = new Bath({
       ...req.body,
     });
     // Save the bath record
     const savedBath = await bath.save();
-
     // Send a success message as a JSON response
-    res.status(201).json({ message: "Baignade enregistré !", bath: savedBath });
+    res.status(201).json({
+      message: "La baignade a correctement été enregistrée.",
+      bath: savedBath,
+    });
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
 /**
  * @function modifyBath
  * @async
- * @description Modifies an existing Bath record in the database.
+ * @description Modifies an existing Bath record in the database and returns the updated record.
  * @see {@link module:BathRoutes} - This function is used in the PUT /api/bath/:id route.
  * @param {Object} req - Express request object, containing the Bath ID in `req.params.id` and updated details in `req.body`.
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.put("/api/bath/:id", [authJwt.verifyToken], controller.modifyBath);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if a required field is missing.
+ * @throws {BadRequest} JSON response with a 400 status if the user ID is invalid.
+ * @throws {NotFound} JSON response with a 404 status if the bath is not found.
+ * @throws {NotFound} JSON response with a 404 status if the user is not found.
+ * @example app.put("/api/bath/:id", [authJwt.verifyToken], controller.modifyBath);
  */
 exports.modifyBath = async (req, res, next) => {
   try {
-    // Verify if the user ID exists
-    const userId = req.body.author;
-    const userExists = await User.findById(userId);
+    // Verify if all required fields are present
+    for (const field of requiredFields) {
+      if (!req.body.hasOwnProperty(field)) {
+        return res.status(400).json({ message: "Champs requis manquants." });
+      }
+    }
 
+    const userId = req.body.author;
+    // Check if userId is valid
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
+    // Verify if the user ID exists
+    const userExists = await User.findById(userId);
     if (!userExists) {
-      // Return a generic error message or a misleading success message
-      return res.status(400).json({ message: "Une erreur s'est produite." });
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
+    }
+    // Verify if the bath ID exists
+    const bathExists = await Bath.findById(req.params.id);
+    if (!bathExists) {
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
     }
     // Update the bath record
     await Bath.updateOne(
@@ -101,8 +142,11 @@ exports.modifyBath = async (req, res, next) => {
     // Send a success message as a JSON response
     res.status(200).json({ message: "Baignade edité !", bath: updatedBath });
   } catch (error) {
+    if (error.name === "CastError") {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
@@ -115,21 +159,29 @@ exports.modifyBath = async (req, res, next) => {
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if the bath ID is invalid.
+ * @throws {BadRequest} JSON response with a 400 status if the user ID is invalid.
  * @throws {NotFound} JSON response with a 404 status if the bath is not found.
- * @example
- * // Route definition in another file
- * app.delete("/api/bath/:id", [authJwt.verifyToken], controller.deleteBath);
+ * @throws {NotFound} JSON response with a 404 status if the user is not found.
+ * @example app.delete("/api/bath/:id", [authJwt.verifyToken], controller.deleteBath);
  */
 exports.deleteBath = async (req, res, next) => {
-  try {
-    // Verify if the user ID exists
-    const userId = req.body.author;
-    const userExists = await User.findById(userId);
+  // Verify if the bath ID is valid
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: "Baignade non trouvée." });
+  }
 
+  try {
+    const userId = req.body.author;
+    // Check if userId is valid
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
+    const userExists = await User.findById(userId);
+    // Verify if the user ID exists
     if (!userExists) {
-      // Return a generic error message or a misleading success message
-      return res.status(400).json({ message: "Une erreur s'est produite." });
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
     }
     // Delete the bath record and get the deleted record
     const deletedBath = await Bath.findOneAndDelete({ _id: req.params.id });
@@ -137,12 +189,11 @@ exports.deleteBath = async (req, res, next) => {
     if (!deletedBath) {
       return res.status(404).json({ message: "Baignade non trouvée." });
     }
-
     // Send a success message as a JSON response
     res.status(200).json({ message: "Baignade supprimé !", bath: deletedBath });
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
@@ -155,10 +206,8 @@ exports.deleteBath = async (req, res, next) => {
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.get("/api/bath", controller.getAllBaths);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @example app.get("/api/bath", controller.getAllBaths);
  */
 exports.getAllBaths = async (req, res, next) => {
   try {
@@ -166,12 +215,11 @@ exports.getAllBaths = async (req, res, next) => {
     const baths = await Bath.find()
       .sort(sortOptions)
       .populate("author", "username");
-
     // Send the fetched baths as a JSON response
     res.status(200).json(baths);
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
@@ -184,29 +232,31 @@ exports.getAllBaths = async (req, res, next) => {
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.get("/api/bath/:id", controller.getOneBath);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if the bath ID is invalid.
+ * @throws {NotFound} JSON response with a 404 status if the bath is not found.
+ * @example app.get("/api/bath/:id", controller.getOneBath);
  */
 exports.getOneBath = async (req, res, next) => {
   try {
+    // Check if the bath ID is valid
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
     // Fetch a single bath by ID
     const bath = await Bath.findOne({ _id: req.params.id }).populate(
       "author",
       "username"
     );
-
     // If the bath is not found, return a 404 error
     if (!bath) {
-      return res.status(404).json({ message: "Bath not found" });
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
     }
-
     // Send the fetched bath as a JSON response
     res.status(200).json(bath);
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
@@ -219,14 +269,23 @@ exports.getOneBath = async (req, res, next) => {
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with the fetched baths.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.get("/api/bath/recent/:limit", controller.getRecentBaths);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if the limit is missing.
+ * @throws {NotFound} JSON response with a 400 status if the limit is invalid.
+ * @example app.get("/api/bath/recent/:limit", controller.getRecentBaths);
  */
 exports.getRecentBaths = async (req, res, next) => {
   try {
+    // check if limit exists
+    if (!req.params.limit) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
+    // Get the limit from the request parameters
     const limit = Number(req.params.limit);
+    // Check if the limit is valid
+    if (isNaN(limit) || limit !== 6) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
     // Fetch recent baths
     const baths = await Bath.find()
       .limit(limit)
@@ -237,7 +296,7 @@ exports.getRecentBaths = async (req, res, next) => {
     res.status(200).json(baths);
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
 
@@ -250,14 +309,25 @@ exports.getRecentBaths = async (req, res, next) => {
  * @param {Object} res - Express response object.
  * @param {function} next - Express next middleware function.
  * @returns {Object} JSON response with a success message.
- * @throws {BadRequest} JSON response with a 400 status if an error occurs.
- * @example
- * // Route definition in another file
- * app.get("/api/bath/recent/:limit", controller.getRecentBaths);
+ * @throws {BadRequest} JSON response with a 500 status if an error occurs.
+ * @throws {BadRequest} JSON response with a 400 status if the user ID is invalid.
+ * @throws {NotFound} JSON response with a 404 status if the user is not found.
+ * @example app.get("/api/bath/recent/:limit", controller.getRecentBaths);
  */
 exports.getAllBathsByUser = async (req, res, next) => {
   try {
+    // Check if userId is valid
+    if (!mongoose.Types.ObjectId.isValid(req.params.userId)) {
+      return res.status(400).json({ message: errorMessages.GENERIC_ERROR });
+    }
+
     const userId = req.params.userId;
+
+    const userExists = await User.findById(userId);
+    // Verify if the user ID exists
+    if (!userExists) {
+      return res.status(404).json({ message: errorMessages.GENERIC_ERROR });
+    }
     // Fetch baths belonging to a single user
     const baths = await Bath.find({ author: userId })
       .sort(sortOptions)
@@ -267,6 +337,6 @@ exports.getAllBathsByUser = async (req, res, next) => {
     res.status(200).json(baths);
   } catch (error) {
     // console.log("Caught an error:", error); // Debug log
-    res.status(400).json({ error });
+    res.status(500).json({ message: errorMessages.INTERNAL_SERVER_ERROR });
   }
 };
